requirements.txt
Pillow==8.0.1
pytesseract==0.3.8
pytz==2021.3
requests==2.27.1
Telethon==1.24.0
Unidecode==1.3.4
jellyfish==0.9.0
python-decouple==3.6